Make single reusable MockAppender class
