+++ test/framework/src/main/java/org/elasticsearch/common/logging/MockAppender.java
@@ -1,17 +1,22 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-package org.elasticsearch.xpack.ml.test;
+package org.elasticsearch.common.logging;
 
 import org.apache.logging.log4j.core.LogEvent;
 import org.apache.logging.log4j.core.appender.AbstractAppender;
 import org.apache.logging.log4j.core.filter.RegexFilter;
 import org.apache.logging.log4j.message.Message;
 
+/**
+ * Mock log appender class that will capture the last log event.
+ */
 public class MockAppender extends AbstractAppender {
     public LogEvent lastEvent;
 
@@ -24,8 +29,12 @@ public class MockAppender extends AbstractAppender {
         lastEvent = event.toImmutable();
     }
 
-    Message lastMessage() {
+    public Message lastMessage() {
         return lastEvent.getMessage();
+    }
+
+    public LogEvent lastEvent() {
+        return lastEvent;
     }
 
     public LogEvent getLastEventAndReset() {
